% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firms.R
\docType{data}
\name{firms}
\alias{firms}
\title{Sample data for firm dynamics}
\format{A data frame with the following columns:
\describe{
\item{tfp}{Firm productivity (or firm performance indicators)}
\item{s}{Market shares occupied the firm.}
\item{id}{A firm identity column.}
\item{t}{A time period column}
}}
\usage{
firms
}
\description{
Firm productivity and market shares adopted to
demonstrate the decomposition of firm dynamics
}
\references{
Zeng, S. (2019). Frontier firms, inefficiency and productivity
dynamics. Presented in EMG Workshop 2019, Sydney.
}
\keyword{datasets}
