% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\docType{data}
\name{mining}
\alias{mining}
\title{Sample Data for Value Added Decomposition}
\format{A data frame with the following columns:
\describe{
\item{year}{A time period column.}
\item{p2}{Output prices.}
\item{w2}{Wages of labour inputs.}
\item{u2}{Prices of capital services.}
\item{y2}{Output quantities.}
\item{h2}{Labour input quantities.}
\item{x2}{Capital services quantities.}
}}
\usage{
mining
}
\description{
Mining inputs and outputs adopted to
demonstrate the decomposition of value added growth.
}
\references{
Zeng, S., Parsons, S., Diewert, W. E. and Fox, K. J. (2018).
Industry and state level value added and productivity
decompositions. Presented in EMG Worshop 2018, Sydney.
}
\keyword{datasets}
