% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_index.R
\name{pm_index}
\alias{pm_index}
\title{Multilateral Price Indexes}
\usage{
pm_index(df, p, qty, id, tm, typ = "geks", len = NULL, lnk = NULL, bsk = "flx", wd = "flx")
}
\arguments{
\item{df}{A data frame sorted by the time period column.}

\item{p}{A string indicating the price column.}

\item{qty}{A string indicating the quantity column.}

\item{id}{A string indicating the identity column.}

\item{tm}{A string indicating the time period column. Each period must
contain two observations at least.}

\item{typ}{Relevant types of price indexes. Options include "geks"
for GEKS price indexes (by default), "ccdi" for CCDI price indexes,
"wtpd" for the weighted time product dummy method, and "gk" for the
Geary-Khamis method.}

\item{len}{Window length for linked indexes using rolling windows. A
single window is set as NULL (by default).}

\item{lnk}{Linking position in rolling windows, effective when `len`
is not NULL. If no linking position is provided, it should be set as NULL
(by default). Other options include "mean" for mean splices and numbers
for specific cases.}

\item{bsk}{The choice of baskets when items are not matched over
multiple periods. Options include "flx" (by default) for a flexible basket
that varies depending on the maximal number of matched items in
two periods each time, and "cst" for a constant basket
that takes the maximal number of matched items across all periods.}

\item{wd}{The choice of windows when items are not matched over multiple
windows. Options include "flx" (by default) for a flexible window that allows for
different items in two windows each time, and "cst" for a constant window
that takes the maximal number of matched items across all windows.}
}
\value{
A data frame consisting of the time period and price indexes.
}
\description{
Multilateral indexes refer to the case when more than two periods
are compared each time. `pm_index()` computes price indexes in a
multilateral approach.
}
