% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_decom.R
\name{roll_div}
\alias{roll_div}
\title{Converting Level Values to Growth Values}
\usage{
roll_div(x)
}
\arguments{
\item{x}{A vector with level values.}
}
\value{
A vector of growth values.
}
\description{
\code{roll_div()} converts level values to growth values for a vector.
}
\examples{
table2 <- value_decom(c("h2","x2"), c("w2","u2"), "y2", "p2", "year", mining)[[2]]
roll_div(table2[, "TFP"])
}
